+++ src/Network/Network/FirewallPolicy/FirewallCustomRule/AzureApplicationGatewayFirewallCustomRuleBase.cs
@@ -35,9 +35,23 @@ namespace Microsoft.Azure.Commands.Network
         public int Priority { get; set; }
 
         [Parameter(
+            Mandatory = false,
+            HelpMessage = "Duration over which Rate Limit policy will be applied. Applies only when ruleType is RateLimitRule.")]
+        [ValidateSet("Onemin", "FiveMins", IgnoreCase = true)]
+        [ValidateNotNullOrEmpty]
+        public string RateLimitDuration { get; set; }
+
+        [Parameter(
+            Mandatory = false,
+            HelpMessage = "Rate Limit threshold to apply in case ruleType is RateLimitRule. Must be greater than or equal to 1")]
+        [ValidateRange(1, 5000)]
+        [ValidateNotNullOrEmpty]
+        public int RateLimitThreshold { get; set; }
+
+        [Parameter(
             Mandatory = true,
             HelpMessage = "Describes type of rule.")]
-        [ValidateSet("MatchRule", IgnoreCase = true)]
+        [ValidateSet("MatchRule", "RateLimitRule", IgnoreCase = true)]
         [ValidateNotNullOrEmpty]
         public string RuleType { get; set; }
 
@@ -46,6 +60,13 @@ namespace Microsoft.Azure.Commands.Network
             HelpMessage = "List of match conditions.")]
         [ValidateNotNullOrEmpty]
         public PSApplicationGatewayFirewallCondition[] MatchCondition { get; set; }
+
+        [Parameter(
+            Mandatory = false,
+            HelpMessage = "Define user session identifier group by clauses.")]
+        [ValidateCount(1, 1)]
+        [ValidateNotNullOrEmpty]
+        public PSApplicationGatewayGroupByUserSession GroupByUserSession { get; set; }
 
         [Parameter(
             Mandatory = true,
@@ -77,7 +98,10 @@ namespace Microsoft.Azure.Commands.Network
                 Name = this.Name,
                 Priority = this.Priority,
                 RuleType = this.RuleType,
+                RateLimitDuration = this.RateLimitDuration,
+                RateLimitThreshold = this.RateLimitThreshold,
                 MatchConditions = this.MatchCondition?.ToList(),
+                GroupByUserSession = this.GroupByUserSession,
                 Action = this.Action,
                 State = this.State
             };